placed work order server requests in seperate file for organization
